client: move secure auth token functions to seperate import path.
